Creating C++ Code For FPP and DoorSwing
Creating a C++ code for third person character movement and capability to fire a door action button that opens the door.

Creating a C++ code for DoorSwing to open door when the character presses a button inside the trigger with both left/right swing

Adding Box Collision to AirBox,NormalDoor and base ground plane connecting.

Also creating a blueprint based of the gamemode and character class and assigning the default pawn as the  BP_MyCharacter.